requirements.txt
atlassian-python-api
pandas
requests
beautifulsoup4
openpyxl
lxml